renamed custom_bias -> custom_data
